Move database config into main.py
