+++ src/card_genius/models.py
@@ -9,21 +9,10 @@ class Contact(SQLModel, table=True):
     phone: str
     email: str
 
-sqlite_file_name = "database.db"
-sqlite_url = f"sqlite:///{sqlite_file_name}"
-
-connect_args = {"check_same_thread": False}
-engine = create_engine(sqlite_url, echo=True, connect_args=connect_args)
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
 
 app = FastAPI()
 
 
-@app.on_event("startup")
-def on_startup():
-    create_db_and_tables()
 
 
 @app.post("/contacts/")
